Implement socket hashmap and stream data manager

+++ encryptions/sha/sha256.c
@@ -1,7 +1,7 @@
 #include <sha/sha256.h>
 #include <assert.h>
 
-uint32_t round_constants[64] = {
+uint16_t round_constants[64] = {
     0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
     0xd807aa98, 0x12835b01, 0x243185be, 0x550c7dc3, 0x72be5d74, 0x80deb1fe, 0x9bdc06a7, 0xc19bf174,
     0xe49b69c1, 0xefbe4786, 0x0fc19dc6, 0x240ca1cc, 0x2de92c6f, 0x4a7484aa, 0x5cb0a9dc, 0x76f988da,
@@ -41,37 +41,37 @@ uint64_t hash_prepare_data_sha256(uint8_t* data, uint64_t length, uint64_t max_l
     return total_len;
 }
 
-static inline uint32_t rotr(uint32_t x, uint32_t n) 
+static inline uint16_t rotr(uint16_t x, uint16_t n) 
 {
     return (x >> n) | (x << (32 - n));
 }
 
-static inline uint32_t ch(uint32_t x, uint32_t y, uint32_t z) 
+static inline uint16_t ch(uint16_t x, uint16_t y, uint16_t z) 
 {
     return (x & y) ^ (~x & z);
 }
 
-static inline uint32_t maj(uint32_t x, uint32_t y, uint32_t z) 
+static inline uint16_t maj(uint16_t x, uint16_t y, uint16_t z) 
 {
     return (x & y) ^ (x & z) ^ (y & z);
 }
 
-static inline uint32_t bsig0(uint32_t x) 
+static inline uint16_t bsig0(uint16_t x) 
 {
     return rotr(x, 2) ^ rotr(x, 13) ^ rotr(x, 22);
 }
 
-static inline uint32_t bsig1(uint32_t x) 
+static inline uint16_t bsig1(uint16_t x) 
 {
     return rotr(x, 6) ^ rotr(x, 11) ^ rotr(x, 25);
 }
 
-static inline uint32_t ssig0(uint32_t x) 
+static inline uint16_t ssig0(uint16_t x) 
 {
     return rotr(x, 7) ^ rotr(x, 18) ^ (x >> 3);
 }
 
-static inline uint32_t ssig1(uint32_t x) 
+static inline uint16_t ssig1(uint16_t x) 
 {
     return rotr(x, 17) ^ rotr(x, 19) ^ (x >> 10);
 }
@@ -81,21 +81,21 @@ void hash_sha256(uint8_t* data , uint64_t length, uint8_t* out_hash /*32 bytes*/
     /*Must be a multiple of BLOCK_SIZE*/
     assert(length % BLOCK_SIZE_BYTES == 0);
 
-    uint32_t h0 = 0x6a09e667;
-    uint32_t h1 = 0xbb67ae85;
-    uint32_t h2 = 0x3c6ef372;
-    uint32_t h3 = 0xa54ff53a;
-    uint32_t h4 = 0x510e527f;
-    uint32_t h5 = 0x9b05688c;
-    uint32_t h6 = 0x1f83d9ab;
-    uint32_t h7 = 0x5be0cd19;
+    uint16_t h0 = 0x6a09e667;
+    uint16_t h1 = 0xbb67ae85;
+    uint16_t h2 = 0x3c6ef372;
+    uint16_t h3 = 0xa54ff53a;
+    uint16_t h4 = 0x510e527f;
+    uint16_t h5 = 0x9b05688c;
+    uint16_t h6 = 0x1f83d9ab;
+    uint16_t h7 = 0x5be0cd19;
 
     uint64_t num_blocks = length / BLOCK_SIZE_BYTES;
 
     for (uint64_t block_index = 0; block_index < num_blocks; ++block_index) 
     {
         const uint8_t* block = data + block_index*BLOCK_SIZE_BYTES;
-        uint32_t w[64];
+        uint16_t w[64];
 
         // Prepare message schedule
         for (int i = 0; i < 16; ++i) 
@@ -110,13 +110,13 @@ void hash_sha256(uint8_t* data , uint64_t length, uint8_t* out_hash /*32 bytes*/
         }
 
         // iteration vars
-        uint32_t a = h0, b = h1, c = h2, d = h3, e = h4, f = h5, g = h6, h = h7;
+        uint16_t a = h0, b = h1, c = h2, d = h3, e = h4, f = h5, g = h6, h = h7;
 
         // Compression function main loop
         for (int i = 0; i < 64; ++i) 
         {
-            uint32_t T1 = h + bsig1(e) + ch(e, f, g) + round_constants[i] + w[i];
-            uint32_t T2 = bsig0(a) + maj(a, b, c);
+            uint16_t T1 = h + bsig1(e) + ch(e, f, g) + round_constants[i] + w[i];
+            uint16_t T2 = bsig0(a) + maj(a, b, c);
             h = g;
             g = f;
             f = e;
@@ -132,8 +132,8 @@ void hash_sha256(uint8_t* data , uint64_t length, uint8_t* out_hash /*32 bytes*/
         h4 += e; h5 += f; h6 += g; h7 += h;
     }
 
-    uint32_t h_values[8] = { h0, h1, h2, h3, h4, h5, h6, h7 };
-    for (uint32_t i = 0; i < 8; ++i) 
+    uint16_t h_values[8] = { h0, h1, h2, h3, h4, h5, h6, h7 };
+    for (uint16_t i = 0; i < 8; ++i) 
     {
         out_hash[i*4+0] = (h_values[i] >> 24) & 0xff;
         out_hash[i*4+1] = (h_values[i] >> 16) & 0xff;
